solar_power_manager/requirements.txt
paho-mqtt
influxdb
pyserial
python-crontab
